Refactor `SoftwareDiagnosticsServerCluster` to remove `SoftwareDiagnosticsLogic` dependency
