+++ src/app/clusters/software-diagnostics-server/BUILD.gn
@@ -30,8 +30,6 @@ source_set("software-diagnostics-server") {
   sources = [
     "SoftwareDiagnosticsCluster.cpp",
     "SoftwareDiagnosticsCluster.h",
-    "SoftwareDiagnosticsLogic.cpp",
-    "SoftwareDiagnosticsLogic.h",
   ]
 
   public_deps = [
